fix pressure var

+++ include/display_api.h
@@ -25,7 +25,7 @@ void ICACHE_FLASH_ATTR
 display_wind_gust(uint8_t wind_gust);
 
 void ICACHE_FLASH_ATTR 
-display_pressure_mm(uint8_t pressure_mm);
+display_pressure_mm(uint16_t pressure_mm);
 
 void ICACHE_FLASH_ATTR 
 display_humidity(uint8_t humidity);
